requirements_climate.txt
# 气候预测系统依赖包
# Climate Prediction System Requirements

# 核心数据处理
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习框架
scikit-learn>=1.0.0
xgboost>=1.5.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 气象数据处理
xarray>=0.19.0
cfgrib>=0.9.9
eccodes>=1.4.0
netcdf4>=1.5.7

# 数据库连接
sqlalchemy>=1.4.0
pymysql>=1.0.0
python-dotenv>=0.19.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 日期时间处理
python-dateutil>=2.8.0
pytz>=2021.1

# 数据验证
jsonschema>=3.2.0

# 进度条
tqdm>=4.62.0

# 日志和配置
pyyaml>=5.4.0
coloredlogs>=15.0

# 性能优化
numba>=0.54.0
joblib>=1.0.0

# 统计分析
statsmodels>=0.12.0

# 文件处理
h5py>=3.3.0
pickle5>=0.0.11

# 网络请求（如果需要下载数据）
requests>=2.25.0
urllib3>=1.26.0

# 开发和测试工具
pytest>=6.2.0
pytest-cov>=2.12.0
flake8>=3.9.0
black>=21.6.0

# Jupyter支持（可选）
jupyter>=1.0.0
ipykernel>=6.0.0

# 内存优化
psutil>=5.8.0

# 并行处理
multiprocessing-logging>=0.3.0